+++ Samples~/ShapesApp/Scripts/ShapesApp.cs
@@ -605,7 +605,7 @@ public class ShapesApp : MonoBehaviour
         {
             if (IsApproovEnabled())
             {
-                approovSendRoutine = request.SendApproovWebRequest();
+                approovSendRoutine = ApproovService.SendWebRequest(request);
             }
             else
             {